I init standalone libs for getTransferLog and getTransferSingleLog.
