+++ lib/viem/decodeEventLog.js
@@ -3,6 +3,8 @@ import {
   zoraCreator1155ImplABI,
 } from "@zoralabs/protocol-deployments";
 import { parseEventLogs } from "viem";
+import getTransferSingleLog from "./getTransferSingleLog";
+import getTransferLog from "./getTransferLog";
 
 const decodeEventLog = (event) => {
   let parsedLog;
@@ -29,45 +31,3 @@ const decodeEventLog = (event) => {
 };
 
 export default decodeEventLog;
-
-const getTransferLog = (logs) => {
-  const transferLogs = parseEventLogs({
-    abi: [
-      {
-        name: "Transfer",
-        type: "event",
-        inputs: [
-          { indexed: true, name: "from", type: "address" },
-          { indexed: true, name: "to", type: "address" },
-          { indexed: true, name: "tokenId", type: "uint256" },
-        ],
-      },
-    ],
-    eventName: "Transfer",
-    logs,
-  });
-  try {
-    console.log("transferLogs:", transferLogs);
-    const { tokenId, to } = transferLogs[0].args;
-    return { collector: to, tokenId };
-  } catch (error) {
-    console.log("No TransferSingle logs found");
-    return {};
-  }
-};
-
-const getTransferSingleLog = (logs) => {
-  const collectionLogs = parseEventLogs({
-    abi: zoraCreator1155ImplABI,
-    eventName: "TransferSingle",
-    logs,
-  });
-  console.log("collectionLogs:", collectionLogs);
-  try {
-    const { to, id, value } = collectionLogs[0].args;
-    return { collector: to, tokenId: id, quantity: value };
-  } catch (error) {
-    console.log("No TransferSingle logs found");
-    return {};
-  }
-};
