+++ lib/viem/getTransferSingleLog.js
@@ -1,0 +1,17 @@
+const getTransferSingleLog = (logs) => {
+  const collectionLogs = parseEventLogs({
+    abi: zoraCreator1155ImplABI,
+    eventName: "TransferSingle",
+    logs,
+  });
+  console.log("collectionLogs:", collectionLogs);
+  try {
+    const { to, id, value } = collectionLogs[0].args;
+    return { collector: to, tokenId: id, quantity: value };
+  } catch (error) {
+    console.log("No TransferSingle logs found");
+    return {};
+  }
+};
+
+export default getTransferSingleLog;
